Refactor protocol/client
The Callisto client no longer assumes the implementation of the OPRF. In
reality, a client only cares about computing P-Hat values. This makes
the code cleaner because in the future if we need to swap out for a
different OPRF construction, we can hide it through this new interface of
computing P-Hat values.

+ Create PHatComputer struct and interface. The struct currently
  encapsulates the V-OPRF (IETF draft) implementation--both the client
  and server/evaluator. The interface on the other hand has no
  requirement of being an OPRF; it simply requires a computer that can
  compute P-Hats.