Add userId to callisto tuple type
We need to store userIDs with callisto tuples. Remember, these IDs are
non-identifying. In this implementation, they're random GUIDs.

We need to add this distinguishing property so that the matching
algorithm can return matches only when there are >= 2 matches on the
same pi value, _from different user IDs_. If the same user reports on
the same perp twice, this is not considered a match as we won't be able
to derive the k value (shamir requires two points).

+ Update callisto tuple tests to check for this new property