Matching no longer needs to return unique user id entries
In an earlier commit, matching was refactored to return a list of
matched pi entries with unique user ids. This list was fed into dloc
decryption so that interpolation wouldn't break (see earlier commit for
details).

A simpler fix is to remove duplicate entries/shares at a lower level before
interpolation. This simplifies a lot of the code at higher levels.

+ Add more tests to check this logic. See shamir_test:TestFindShamirKValue_CorrectlyInterpolatesWhenMixed